+++ CHANGELOG.md
@@ -2,3 +2,10 @@
 + Added a basic overlay template with comments
 * Fixed lost hold stealth bonus being triggered by non-hold guards
 * Rounded health values from decimal values to integers
+
+### H3Status 0.2.0
++ Added `TNHLevelEvent`
++ Added `TNHTokenEvent`
++ Added game & mod version info
+* Merged `playerDamage`, `playerHeal`, and `playerKill` into `healthEvent`
+* Renamed `playerBuff` to `buffEvent`